api/requirements.txt
# TelegramOllama API Dependencies
# FastAPI сервер для Web UI и bot API

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Data Validation & Models
# COMPATIBLE VERSION: aiogram 3.3.0 requires pydantic <2.6
# langchain requires pydantic >=2.7.4 — NOT USED in API
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-settings==2.1.0

# Auth
pyjwt==2.8.0
# bcrypt is used directly in auth.py (not via passlib)
# Version 5.0.0 is stable on Linux/Windows and used in production
bcrypt==5.0.0
# passlib kept for potential future migration but NOT actively used
passlib==1.7.4

# Utils
python-dateutil==2.8.2
pytz==2024.1

# OCR (Phase 13+)
pytesseract>=0.3.10
Pillow>=10.0.0

# CORS
# (already in fastapi)

# NOTE: aiogram and langchain are ONLY for bot/agent containers
# API does NOT use them directly
